// client/src/types/IProductParamsId.ts
export interface IProductParamsId {
  imageUrl: string;
  images: string;
  name: string;
  price: number | string;
  brand: string;
  memory: number | string;
  ram: number | string;
  cpuCores: number | string;
  screenSize: number | string;
  batteryCapacity: number | string;
  color: string;
  productCode: number | string;
  description: string;
  shortDesc: string;
  id: any;
}